Move EventHeader and EventLoggingTypes into MessageDef because that seems to be the pimary use
