+++ examples/chip-tool/commands/clusters/DataModelLogger.h
@@ -23,7 +23,7 @@
 #include <app-common/zap-generated/cluster-objects.h>
 #include <app/ConcreteAttributePath.h>
 #include <app/ConcreteCommandPath.h>
-#include <app/EventHeader.h>
+#include <app/MessageDef/EventHeader.h>
 #include <app/MessageDef/StatusIB.h>
 #include <app/data-model/DecodableList.h>
 #include <commands/common/RemoteDataModelLogger.h>
