+++ src/app/tests/TestEventOverflow.cpp
@@ -23,7 +23,7 @@
  */
 
 #include <app/EventLoggingDelegate.h>
-#include <app/EventLoggingTypes.h>
+#include <app/MessageDef/EventLoggingTypes.h>
 #include <app/EventManagement.h>
 #include <app/InteractionModelEngine.h>
 #include <app/tests/AppTestContext.h>
