+++ src/app/tests/integration/MockEvents.cpp
@@ -25,7 +25,7 @@
 
 #include "MockEvents.h"
 #include "common.h"
-#include <app/EventLoggingTypes.h>
+#include <app/MessageDef/EventLoggingTypes.h>
 #include <app/EventManagement.h>
 #include <lib/core/ErrorStr.h>
 #include <platform/CHIPDeviceLayer.h>
