+++ src/app/zap-templates/templates/app/cluster-objects.zapt
@@ -8,7 +8,7 @@
 #include <app/data-model/List.h>
 #include <app/data-model/NullObject.h>
 #include <app/ConcreteAttributePath.h>
-#include <app/EventLoggingTypes.h>
+#include <app/MessageDef/EventLoggingTypes.h>
 #include <app/util/basic-types.h>
 #include <lib/core/ClusterEnums.h>
 #include <lib/support/BitMask.h>
