+++ src/controller/java/AndroidCallbacks.cpp
@@ -21,7 +21,7 @@
 #include <controller/java/CHIPAttributeTLVValueDecoder.h>
 #include <controller/java/CHIPEventTLVValueDecoder.h>
 #endif
-#include <app/EventLoggingTypes.h>
+#include <app/MessageDef/EventLoggingTypes.h>
 #include <jni.h>
 #include <lib/core/ErrorStr.h>
 #include <lib/support/CHIPJNIError.h>
