+++ src/darwin/Framework/CHIP/MTRBaseDevice_Internal.h
@@ -23,8 +23,8 @@
 #include <app/ConcreteCommandPath.h>
 #include <app/ConcreteEventPath.h>
 #include <app/DeviceProxy.h>
-#include <app/EventHeader.h>
-#include <app/EventLoggingTypes.h>
+#include <app/MessageDef/EventHeader.h>
+#include <app/MessageDef/EventLoggingTypes.h>
 #include <app/EventPathParams.h>
 #include <lib/core/CHIPVendorIdentifiers.hpp>
 #include <lib/core/TLVTags.h>
