+++ src/darwin/Framework/CHIP/MTREventTLVValueDecoder_Internal.h
@@ -21,7 +21,7 @@
 #import <Foundation/Foundation.h>
 
 #include <app/ConcreteAttributePath.h>
-#include <app/EventLoggingTypes.h>
+#include <app/MessageDef/EventLoggingTypes.h>
 #include <lib/core/CHIPError.h>
 #include <lib/core/TLV.h>
