+++ src/darwin/Framework/CHIP/templates/MTREventTLVValueDecoder-src.zapt
@@ -12,8 +12,8 @@
 #include <app-common/zap-generated/ids/Attributes.h>
 #include <app-common/zap-generated/ids/Events.h>
 #include <app-common/zap-generated/ids/Clusters.h>
-#include <app/EventHeader.h>
-#include <app/EventLoggingTypes.h>
+#include <app/MessageDef/EventHeader.h>
+#include <app/MessageDef/EventLoggingTypes.h>
 #include <lib/support/TypeTraits.h>
 
 using namespace chip;
